New version of the SignalProcessor.
Blocks are processed anew everytime before they can be rendered.
Both rendering and processing is in the main thread. Synchronization
between OpenCL and OpenGL is through blocking calls clFinish and glFinish.

Program now has one cache on the GPU that holds raw data blocks.

Loading to the cache is performed in a worker thread. Synchronization
between this thread and the main thread is done via OpenCL events.

In this version filtering is broken.


Former-commit-id: 78e596994a50b5cc610b1e469b23344682568074
Former-commit-id: 0ad54941b7b6739404a8a0ae8eecb43e5b508731