Fix the GPU memory leak.
The problem was caused by clReleaseContext not releasing all resources.
This slowly ate all the available memory, and then the program crashed with
an out of memory exception.

The problem was solved by limiting the number of created context to a maximum
of 2 at once. This makes the leak insignificant, and the program is more stable now.


Former-commit-id: a021a4aea2153e8ddcdb56103ab63cd48f176f67
Former-commit-id: 9fba057e7e73da659f25d092b3a80b52200d0686