Change how position is tracked.
Now the global position is in samples from the start of data.
Working  with position is now trivial as there is no need to
convert it every time from "virtual" position using the virtual width.
Position and position indicator now work together

These are the only use-cases for virtual width and position now:

* the main scroll bar to adjust the page and step size
* when setting seconds per page
* Canvas for rendering

Also move Sync functionality out of the main window. The change in position
representation means Sync in not compatible with previous versions.